techpack: camera: isp: Change logic of acquiring dual composite group
There is a possibility of composite groups in free dual
composite group list are not in sequential order. In case
of dual ife use case, if composite groups are not in
sequential manner then we landed into a scenario
where two different IFE with same resource have two
different composite group index, which is not expected.

This commit ensures if it is dual IFE usecase then at the time
of acquiring composite group, both IFEs have same composite
group ID for same resource.

CRs-Fixed: 2912721
Change-Id: I734d1d1660c1e70e0480e2dc0b8377a1c441c634
Signed-off-by: Shravya Samala <shravyas@codeaurora.org>